Picture_bot/requirements.txt
requests==2.27.1
opencv-python==4.5.5.62
transliterate==1.10.2
aiogram==2.18
torch~=1.10.2
torchvision~=0.11.3
numpy~=1.21.5
Pillow~=9.0.1
tqdm~=4.62.3